arreglos de setup de hibernate

+++ myApp/persistence/src/main/java/ar/edu/itba/paw/persistence/RoleDaoJdbc.java
@@ -46,12 +46,12 @@ public class RoleDaoJdbc implements RoleDao{
     }
 
     @Override
-    public int addRoleToUser(int roleId, int userId) {
+    public int addRoleToUser(int roleId, Long userId) {
         return jdbcTemplate.update("INSERT INTO userRoles(roleId, userId) VALUES (?, ?)", roleId, userId);
     }
 
     @Override
-    public List<Role> getUserRoles(int userid) {
+    public List<Role> getUserRoles(Long userid) {
         final List<Role> list = jdbcTemplate.query("SELECT r.roleid as roleid, role FROM roles r JOIN userroles u ON r.roleid = u.roleid WHERE userid = ?",
                 new Object[] {userid}, ROW_MAPPER);
         return list.isEmpty() ? new ArrayList<>() : list;
